Factor out layers, wrap some Quil fns
In the process of factoring out layer logic into a new ns, I made a parameterized layer. One of the parameters was a color. Quil doesn't let you make colors until the sketch is running, which makes it hard to parameterize layers. I introduced a vector-based color format and wrapped some Quil fns that take colors to work with the new color representation.