Fix setting background to black on each draw
Was trying to set background to black so that if no layers present, sketch is all black instead of the default grey. Originally, I set this draw, but that overwrote the backgrounds layers set in their setup functions. Moved it to setup, to be executed before any of the layer setup fns.