Display scene display-name, change config format
Changed config format so that scenes is a vector of maps, not a map itself. Each scene now has an :id field (also made this change to regions config format). This way the front-end can see the scene's name in the scene spec directly rather than looking at the whole scenes map.